# apps/api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Automação de Contratos",
    "version": "1.0.0",
    "description": "Documentação da API"
  },
  "servers": [
    {
      "url": "http://localhost:3007"
    }
  ],
  "components": {
    "schemas": {
...
      "Sellers": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "BigInt serializado"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "cpf": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string"
          },
          "head_id": {
            "type": "string",
            "nullable": true,
            "description": "BigInt serializado"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "update_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/sellers/{id}": {
      "get": {
        "summary": "Obter vendedor por ID",
        "tags": [
          "Sellers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sellers"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualizar vendedor",
        "tags": [
          "Sellers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "data": {
                    "$ref": "#/components/schemas/Sellers"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Atualizado"
          }
        }
      },
      "delete": {
        "summary": "Deletar vendedor",
        "tags": [
          "Sellers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletado"
          }
        }
      }
    },
    "/sellers/email/{email}": {
      "get": {
        "summary": "Obter vendedor por email",
        "tags": [
          "Sellers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sellers"
                }
              }
            }
          }
        }
      }
    },
    "/sellers": {
      "get": {
        "summary": "Listar vendedores",
        "tags": [
          "Sellers"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "head_id",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por ID do Head (BigInt serializado)"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por tipo de usuário (seller, head, coord, admin)"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Sellers"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar novo vendedor",
        "tags": [
          "Sellers"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "data": {
                    "$ref": "#/components/schemas/Sellers"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Criado"
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Autenticação de usuário",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login bem-sucedido"
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        }
      }
    }
  },
...
}